+++ unocss/rules/index.js
@@ -1,0 +1,6 @@
+export { propertyMap } from './propertyMap.js'
+export { otherRules } from './otherRules.js'
+export { borderRules } from './borderRules.js'
+export { textDecorationRules } from './textDecorationRules.js'
+export { pseudoElementRules } from './pseudoElement.js'
+export { pseudoClassRules } from './pseudoClass.js'